Making algorithm to make trianle bodies for Terrain OctoTree

+++ OpenGLearning/CoatlPhysicsEngine/src/Collisions/Terrain.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "../OptimizaAlgorithm/OctoTree.h"
+#include "../Shapes/Triangles.h"
 namespace CoatlPhysicsEngine
 {
 	class Terrain
@@ -7,7 +8,7 @@ namespace CoatlPhysicsEngine
 	private:
 		std::unique_ptr<OctoTree> TerWorld;
 	public:
-		Terrain(std::vector<glm::vec3> Mesh, std::vector<int> Index,int Dim);
+		Terrain(std::vector<glm::vec3> Mesh, std::vector<int> Index,float Dim);
 		~Terrain();
 		std::vector<std::shared_ptr<Bodies>> GetTerrs(glm::vec3 Pos, float Ext);
 	};
